Moved FieldHelpers to utils
