+++ signer/Cargo.toml
@@ -9,6 +9,7 @@ path = "src/lib.rs"
 [dependencies]
 oracle = { path = "../oracle" }
 mina-curves = { path = "../curves" }
+o1-utils = { path = "../utils" }
 
 ark-ec = { version = "0.3.0", features = [ "parallel" ] }
 ark-ff = { version = "0.3.0", features = [ "parallel", "asm" ] }
