+++ utils/Cargo.toml
@@ -12,6 +12,7 @@ ark-serialize = "0.3.0"
 rayon = "1.3.0"
 serde = "1.0.130"
 serde_with = "1.10.0"
+hex = "0.4"
 
 [dev-dependencies]
 mina-curves = { path = "../curves" }
