+++ utils/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod dense_polynomial;
 pub mod evaluations;
+pub mod field_helpers;
 pub mod serialization;
 
 pub use dense_polynomial::ExtendedDensePolynomial;
 pub use evaluations::ExtendedEvaluations;
+pub use field_helpers::FieldHelpers;
